+++ main/VRSubprocessHost.cs
@@ -336,10 +336,11 @@ public sealed class VRSubprocessHost : IDisposable
     public void SfReset() => Send("sf_reset");
 
     public void FsConnect(uint leftButton, uint rightButton, string outputDevice, int activationRadius,
-                          uint leftRecordButton, uint rightRecordButton)
+                          uint leftRecordButton, uint rightRecordButton,
+                          int gifMaxDim, int gifFps)
     {
         FsConnected = true;
-        Send("fs_connect", new { leftButton, rightButton, outputDevice, activationRadius, leftRecordButton, rightRecordButton });
+        Send("fs_connect", new { leftButton, rightButton, outputDevice, activationRadius, leftRecordButton, rightRecordButton, gifMaxDim, gifFps });
     }
 
     public void FsDisconnect()
@@ -350,8 +351,9 @@ public sealed class VRSubprocessHost : IDisposable
     }
 
     public void FsConfig(uint leftButton, uint rightButton, int activationRadius,
-                         uint leftRecordButton, uint rightRecordButton)
-        => Send("fs_config", new { leftButton, rightButton, activationRadius, leftRecordButton, rightRecordButton });
+                         uint leftRecordButton, uint rightRecordButton,
+                         int gifMaxDim, int gifFps)
+        => Send("fs_config", new { leftButton, rightButton, activationRadius, leftRecordButton, rightRecordButton, gifMaxDim, gifFps });
 
     public void FsSetOutput(string deviceName)
         => Send("fs_set_output", new { deviceName });
@@ -407,9 +409,9 @@ public sealed class VRSubprocessHost : IDisposable
     public void SfDisconnect() { }
     public void SfConfig(float a, bool b, bool c, bool d, uint e, uint f, uint g, uint h) { }
     public void SfReset() { }
-    public void FsConnect(uint a, uint b, string c, int d, uint e, uint f) { }
+    public void FsConnect(uint a, uint b, string c, int d, uint e, uint f, int g, int h) { }
     public void FsDisconnect() { }
-    public void FsConfig(uint a, uint b, int c, uint d, uint e) { }
+    public void FsConfig(uint a, uint b, int c, uint d, uint e, int f, int g) { }
     public void FsSetOutput(string a) { }
     public void FsGetDevices() { }
     public event System.Action<Newtonsoft.Json.Linq.JObject>? OnFsUpdate;